Factors out models and input pipeline for seq2seq example
