+++ examples/seq2seq/train_test.py
@@ -24,32 +24,40 @@ from jax import random
 import numpy as np
 import optax
 
+import input_pipeline
 import train
+import models
 
 jax.config.parse_flags_with_absl()
 
+def create_ctable(chars = '0123456789+= '):
+  return input_pipeline.CharacterTable(chars)
 
-def create_test_state():
-  model = train.Seq2seq(teacher_force=False, hidden_size=train.FLAGS.hidden_size)
-  params = train.get_initial_params(model, jax.random.PRNGKey(0))
+
+def create_train_state(ctable):
+  model = models.Seq2seq(teacher_force=False, 
+      hidden_size=train.FLAGS.hidden_size, vocab_size=ctable.vocab_size)
+  params = train.get_initial_params(model, jax.random.PRNGKey(0), ctable)
   tx = optax.adam(train.FLAGS.learning_rate)
-  return train_state.TrainState.create(
+  state = train_state.TrainState.create(
       apply_fn=model.apply, params=params, tx=tx)
+  return state
 
 
 class TrainTest(absltest.TestCase):
 
   def test_character_table(self):
+    ctable = create_ctable()
     text = '410+19'
-    enc_text = train.CTABLE.encode(text)
-    dec_text = train.CTABLE.decode(enc_text)
+    enc_text = ctable.encode(text)
+    dec_text = ctable.decode(enc_text)
     # The text is possibly padded with whitespace, but the trimmed output should
     # be equal to the input.
     self.assertEqual(text, dec_text.strip())
 
   def test_mask_sequences(self):
     np.testing.assert_equal(
-        train.mask_sequences(
+        input_pipeline.mask_sequences(
             np.arange(1, 13).reshape((4, 3)),
             np.array([3, 2, 1, 0])
         ),
@@ -66,33 +74,35 @@ class TrainTest(absltest.TestCase):
         functools.partial(jax.nn.one_hot, num_classes=4))(
             np.array([[0, 1, 0], [1, 0, 2], [1, 2, 0], [1, 2, 3]]))
     np.testing.assert_equal(
-        train.get_sequence_lengths(oh_sequence_batch, eos_id=0),
+        input_pipeline.get_sequence_lengths(oh_sequence_batch, eos_id=0),
         np.array([1, 2, 3, 3], np.int32)
     )
     np.testing.assert_equal(
-        train.get_sequence_lengths(oh_sequence_batch, eos_id=1),
+        input_pipeline.get_sequence_lengths(oh_sequence_batch, eos_id=1),
         np.array([2, 1, 1, 1], np.int32)
     )
     np.testing.assert_equal(
-        train.get_sequence_lengths(oh_sequence_batch, eos_id=2),
+        input_pipeline.get_sequence_lengths(oh_sequence_batch, eos_id=2),
         np.array([3, 3, 2, 2], np.int32)
     )
 
   def test_train_one_step(self):
-    batch = train.get_batch(128)
+    ctable = create_ctable()
+    batch = ctable.get_batch(128)
 
-    state = create_test_state()
+    state = create_train_state(ctable)
     key = random.PRNGKey(0)
-    _, train_metrics = train.train_step(state, batch, key)
+    _, train_metrics = train.train_step(state, batch, key, ctable.eos_id)
 
     self.assertLessEqual(train_metrics['loss'], 5)
     self.assertGreaterEqual(train_metrics['accuracy'], 0)
 
   def test_decode_batch(self):
+    ctable = create_ctable()
+    batch = ctable.get_batch(5)
     key = random.PRNGKey(0)
-    state = create_test_state()
-    batch = train.get_batch(5)
-    train.decode_batch(state.params, batch, key)
+    state = create_train_state(ctable)
+    train.decode_batch(state, batch, key, ctable)
 
 if __name__ == '__main__':
   absltest.main()
